Put everything back together again - and it works! (in serial)
