+++ checkpoint.py
@@ -7,34 +7,9 @@ import json
 from . import disk
 from .utilities import message
 
-from . import frames
+from . import frame
 from . import paths
-
-def save_builts(builts, path):
-
-    if uw.mpi.rank == 0:
-
-        builtsDir = os.path.join(path, 'builts')
-        if not os.path.isdir(builtsDir):
-            os.makedirs(builtsDir)
-
-        for builtName, built in sorted(builts.items()):
-            inputs = built.inputs
-            scripts = built.scripts
-            for index, script in enumerate(scripts):
-                tweakedPath = os.path.splitext(script)[0] + ".py"
-                scriptName = builtName + '_' + str(index)
-                newPath = os.path.join(builtsDir, scriptName + ".py")
-                shutil.copyfile(tweakedPath, newPath)
-            inputsFilename = os.path.join(builtsDir, builtName + '.json')
-            with open(inputsFilename, 'w') as file:
-                 json.dump(inputs, file)
-
-def save_json(jsonObj, name, path):
-    if uw.mpi.rank == 0:
-        jsonFilename = os.path.join(path, name + '.json')
-        with open(jsonFilename, 'w') as file:
-             json.dump(jsonObj, file)
+from . import built
 
 class Checkpointer:
 
@@ -48,6 +23,7 @@ class Checkpointer:
             modeltime = None,
             inFrames = [],
             info = {},
+            framescript = None,
             ):
 
         self.figs = figs
@@ -57,8 +33,9 @@ class Checkpointer:
         self.modeltime = modeltime
         self.builts = builts
         self.inFrames = inFrames
-        self.stamps = frames._frametools.make_stamps(self.builts)
+        self.stamps = built.make_stamps(builts)
         self.info = info
+        self.framescript = framescript
 
     def checkpoint(
             self,
@@ -97,11 +74,16 @@ class Checkpointer:
                 if not os.path.isdir(path):
                     os.makedirs(path)
 
-                save_builts(self.builts, path)
+                built.save_builtsDir(self.builts, path)
 
-                save_json(self.stamps, 'stamps', path)
+                disk.save_json(self.stamps, 'stamps', path)
 
-                save_json(self.info, 'info', path)
+                disk.save_json(self.info, 'info', path)
+
+                if not self.framescript is None:
+                    disk.save_script(
+                        self.framescript, 'framescript', path
+                        )
 
             for inFrame in self.inFrames:
                 inFrame.checkpoint(
