+++ initials/extents.py
@@ -2,8 +2,11 @@ from underworld import function as fn
 import numpy as np
 from ._IC import _IC
 
-def build(*args, **kwargs):
-    return IC(*args, **kwargs)
+def build(*args, name = None, **kwargs):
+    built = IC(*args, **kwargs)
+    if type(name) == str:
+        built.name = name
+    return built
 
 class IC(_IC):
