+++ initials/load.py
@@ -4,11 +4,14 @@ import os
 
 from .. import paths
 from .. import mapping
-from .. import frames
+from .. import frame
 from ._IC import _IC
 
-def build(*args, **kwargs):
-    return IC(*args, **kwargs)
+def build(*args, name = None, **kwargs):
+    built = IC(*args, **kwargs)
+    if type(name) == str:
+        built.name = name
+    return built
 
 class IC(_IC):
