+++ observers/_observer.py
@@ -1,5 +1,5 @@
 from ..utilities import check_reqs
-from ..builts import Built
+from ..built import Built
 
 class Observer(Built):
