+++ tests/framestest.py
@@ -1,5 +1,5 @@
 from .. import initials
-from .. import frames
+from .. import frame
 from .. import systems
 from ..utilities import message
 import underworld as uw
